test(project): split test_verifiers_project.py under the bloat-gate limit
746 lines tripped the shared bloat gate's 300-LOC Stop-hook block. Split
the FR-01.02 gate-specific tests (check_basis_forbids_assumed,
check_criteria_free_of_implementation_detail, check_no_empty_split,
check_starting_guidance_present) into two new files by check, and moved
seed_canon_project/_write_grill_trace/_write_splits_config into a shared
_project_check_fixtures.py helper module. All four files now stay under
the 300-line limit; no test behavior changed.
